Fix Memleak

+++ Purring_Engine/src/Input/InputSystem.h
@@ -114,7 +114,7 @@ namespace PE
     private:
         static std::vector<KeyPressedEvent> m_KeyDown; // vector holding current held down buttons
         static std::vector<MouseButtonHoldEvent> m_MouseDown; // vector holding current held down buttons
-        static std::map<int, int> m_KeyTriggered;
+        static std::map<int, float> m_KeyTriggered;
         static float m_bufferTime; // has to be static to be used by the callbacks
     };
 }
